Page Table extract
