+++ uberfire-widgets/uberfire-widgets-table/.gitignore
@@ -1,0 +1,18 @@
+/target
+/local
+
+# Eclipse, Netbeans and IntelliJ files
+/.*
+/**/.*
+!.gitignore
+/nbproject
+*.ipr
+*.iws
+*.iml
+
+# Repository wide ignore mac DS_Store files
+.DS_Store
+
+# Created by Zanata
+/org.uberfire
+
