+++ uberfire-widgets/uberfire-widgets-table/src/main/java/org/uberfire/ext/widgets/table/client/ColumnChangedHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 JBoss, by Red Hat, Inc
+ * Copyright 2016 JBoss, by Red Hat, Inc
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.uberfire.ext.widgets.common.client.tables;
+package org.uberfire.ext.widgets.table.client;
 
 public interface ColumnChangedHandler {
     void beforeColumnChanged();
