+++ uberfire-widgets/uberfire-widgets-table/src/main/java/org/uberfire/ext/widgets/table/client/UberfireSimpleTable.ui.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-  ~ Copyright 2015 JBoss, by Red Hat, Inc
+  ~ Copyright 2016 JBoss, by Red Hat, Inc
   ~
   ~ Licensed under the Apache License, Version 2.0 (the "License");
   ~ you may not use this file except in compliance with the License.
@@ -18,8 +18,7 @@
 <ui:UiBinder xmlns:ui="urn:ui:com.google.gwt.uibinder"
              xmlns:g="urn:import:com.google.gwt.user.client.ui"
              xmlns:b="urn:import:org.gwtbootstrap3.client.ui"
-             xmlns:bg="urn:import:org.gwtbootstrap3.client.ui.gwt"
-             xmlns:uf="urn:import:org.uberfire.ext.widgets.table.client">
+             xmlns:bg="urn:import:org.gwtbootstrap3.client.ui.gwt">
 
   <ui:style>
     .dataGridContainer {
@@ -27,9 +26,8 @@
       padding-bottom: 10px;
     }
 
-    .leftToolBar {
-      float: left;
-      text-align: left
+    .horizontalContainer {
+      height: 32px;
     }
 
     .rightToolBar {
@@ -40,16 +38,12 @@
     .centerToolBar {
       text-align: center
     }
-
-    .pager {
-      margin: 0px;
-    }
   </ui:style>
 
   <g:FlowPanel addStyleNames="{style.dataGridContainer}">
-    <g:HorizontalPanel ui:field="toolbarContainer" width="100%">
+    <g:HorizontalPanel ui:field="toolbarContainer" width="100%" addStyleNames="{style.horizontalContainer}">
       <g:cell width="33%">
-        <g:FlowPanel width="100%" ui:field="leftToolbar" addStyleNames="{style.leftToolBar}"></g:FlowPanel>
+        <g:FlowPanel width="100%" ui:field="leftToolbar"/>
       </g:cell>
       <g:cell width="33%">
         <g:FlowPanel width="100%" ui:field="centerToolbar" addStyleNames="{style.centerToolBar}"></g:FlowPanel>
@@ -57,21 +51,13 @@
       <g:cell width="33%" horizontalAlignment="right">
         <g:HorizontalPanel ui:field="rightToolbar">
           <g:FlowPanel width="100%" ui:field="rightActionsToolbar" addStyleNames="{style.rightToolBar}"/>
-          <b:Button ui:field="columnPickerButton" dataToggle="BUTTON" icon="LIST_UL"/>
+          <b:Button ui:field="columnPickerButton" />
         </g:HorizontalPanel>
       </g:cell>
     </g:HorizontalPanel>
     <g:SimplePanel>
-      <bg:DataGrid ui:field="dataGrid" />
+      <bg:DataGrid ui:field="dataGrid"/>
     </g:SimplePanel>
-    <g:HorizontalPanel width="100%">
-      <g:cell width="50%" horizontalAlignment="left" verticalAlignment="middle">
-        <b:ListBox ui:field="pageSizesSelector" width="100px"/>
-      </g:cell>
-      <g:cell width="50%" horizontalAlignment="right" verticalAlignment="top">
-        <uf:UberfireSimplePager ui:field="pager" addStyleNames="pagination pagination-right pull-right {style.pager}"/>
-      </g:cell>
-    </g:HorizontalPanel>
   </g:FlowPanel>
 
 </ui:UiBinder>
