+++ uberfire-widgets/uberfire-widgets-table/src/main/java/org/uberfire/ext/widgets/table/client/resources/UFTableCss.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2016 JBoss, by Red Hat, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.uberfire.ext.widgets.table.client.resources;
+
+import com.google.gwt.resources.client.CssResource;
+
+public interface UFTableCss
+        extends
+        CssResource {
+
+    @ClassName("dataGrid")
+    String dataGrid();
+
+    @ClassName("dataGridRow")
+    String dataGridRow();
+
+    @ClassName("dataGridHeader")
+    String dataGridHeader();
+
+    @ClassName("dataGridContent")
+    String dataGridContent();
+
+    @ClassName("dataGridMain")
+    String dataGridMain();
+
+    @ClassName("columnPickerButton")
+    String columnPickerButton();
+
+    @ClassName("columnPickerPopup")
+    String columnPickerPopup();;
+}
