+++ uberfire-widgets/uberfire-widgets-table/src/main/resources/META-INF/beans.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<beans xmlns="http://xmlns.jcp.org/xml/ns/javaee" bean-discovery-mode="none">
+  <scan>
+    <exclude name="org.uberfire.ext.widgets.table.client.**"/>
+  </scan>
+</beans>
