+++ uberfire-widgets/uberfire-widgets-table/src/main/resources/org/uberfire/ext/widgets/table/UberfireTableWidget.gwt.xml
@@ -18,11 +18,5 @@
 <!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.4.0//EN"
     "http://google-web-toolkit.googlecode.com/svn/tags/2.5.0/distro-source/core/src/gwt-module.dtd">
 <module>
-  <inherits name="org.gwtbootstrap3.GwtBootstrap3NoTheme"/>
-  <inherits name="org.gwtbootstrap3.extras.datepicker.DatePicker"/>
-
-  <inherits name="org.uberfire.UberfireAPI"/>
-  <inherits name="org.uberfire.ext.services.UberfireWidgetsServicesAPI"/>
-  <inherits name="org.uberfire.ext.widgets.table.UberfireTableWidget"/>
 
 </module>
