+++ uberfire-widgets/uberfire-widgets-table/src/main/resources/org/uberfire/ext/widgets/table/client/resources/css/UberfireSimplePager.css
