+++ uberfire-widgets/uberfire-widgets-table/src/main/resources/org/uberfire/ext/widgets/table/client/resources/css/uftable.css
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2015 JBoss, by Red Hat, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+.dataGrid td div {
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+}
+
+.dataGridRow:first-child td {
+    border-top: none;
+}
+.dataGridHeader {
+
+}
+
+.dataGridHeader thead th {
+    overflow: hidden;
+    text-overflow: ellipsis;
+}
+
+.dataGridContent {
+    border-top: none;
+}
+
+.dataGridMain{
+
+}
+
+.columnPickerButton {
+
+}
+
+.columnPickerPopup {
+    background-color: #FFFFFF;
+    padding: 5px;
+    border: 1px solid #d1d1d1;
+    z-index: 2000;
+}
+
+@external .checkbox;
+
+.columnPickerPopup .checkbox {
+    margin-top: 2px;
+    margin-bottom: 2px;
+}
